\name{disturbances}
\alias{disturbances}
\alias{disturbances.intReg}
\title{
  return the type of disturbancies term used in interval regression
}
\description{
  return the type of disturbancies term used in estimated
  \dQuote{intReg} objects.
}
\usage{
\method{disturbances}{intReg}(object)
}
\arguments{
  \item{object}{
    \dQuote{intReg} object
}
}
\value{
  character string, specifying the type of disturbancies
}
\author{
  Ott Toomet
}
\keyword{methods}
\keyword{utilities}
