\name{disturbancies}
\alias{disturbancies}
\alias{disturbancies.intReg}
\title{
  return the type of disturbancies term used in interval regression
}
\description{
  return the type of disturbancies term used in estimated
  \dQuote{intReg} objects.
}
\usage{
\method{disturbancies}{intReg}(object)
}
\arguments{
  \item{object}{
    \dQuote{intReg} object
}
}
\value{
  character string, specifying the type of disturbancies
}
\author{
  Ott Toomet
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
