\name{model.matrix.intReg}
\alias{model.matrix.intReg}

\title{Design Matrix of Interval Regression Models}

\description{
   Create design matrix of interval regression models
}

\usage{
\method{model.matrix}{intReg}( object, ... )
}

\arguments{
   \item{object}{object of class \code{intReg}.}
   \item{\dots}{currently not used.}
}

\value{
   The design matrix of interval regression models.
}

\author{Arne Henningsen,
   Ott Toomet \email{otoomet@ut.ee}}

\seealso{\code{\link{intReg}},
   \code{\link{model.matrix}}, and \code{\link{model.frame.intReg}}}

\keyword{methods}
