\name{intervals}
\alias{intervals}
\alias{intervals.intReg}
\title{
  return interval boundaries for each observation
}
\description{
  find interval boundaries for estimated interval regression models.
}
\usage{
intervals(object)
}
\arguments{
  \item{object}{
    'intReg' object
}
}
\value{
  a matrix with columns \sQuote{lb} and \sQuote{ub} for lower and upper
  bound respectively.
}
\author{
  Ott Toomet
}
\keyword{methods}
\keyword{utilities}
